moved shared into utils
