+++ src/components/GalleryBackground/animations/Plains.tsx
@@ -1,4 +1,4 @@
-import { getColDarkCss } from './shared';
+import { getColDarkCss } from '../../../utils/shared';
 
 export const PlainBlack = () => (
   <div className='w-full h-full' style={{backgroundColor: getColDarkCss(1)}} />
