+++ src/components/GalleryBackground/animations/Template.tsx
@@ -1,7 +1,7 @@
 import p5 from 'p5';
 import { useP5DupeRemover } from '../../utils/p5DupeRemover';
 import Sketch from 'react-p5';
-import { cols } from './shared';
+import { cols } from '../../../utils/shared';
 
 export const Template = () => {
   const setParent = useP5DupeRemover();
