+++ src/components/GalleryBackground/animations/TwinCircles.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react';
 import { useP5DupeRemover } from '../../utils/p5DupeRemover';
 import Sketch from 'react-p5';
 import { getCurrentBreakpoints } from '../../utils/breakpointUtils';
-import { cols } from './shared';
+import { cols } from '../../../utils/shared';
 
 interface Position {
   x: number;
